Initial timeseries coverage for sensor
