+++ helium_commander/resource.py
@@ -8,7 +8,6 @@ import click
 from builtins import filter as _filter
 from collections import OrderedDict
 from contextlib import contextmanager
-from functools import reduce
 
 
 def filter_mac(id_rep):
@@ -141,36 +140,3 @@ def display_resources(cls, client, resources, **kwargs):
 Resource.display_writer = classmethod(display_writer)
 Resource.display_map = classmethod(display_map)
 Resource.display = classmethod(display_resources)
-
-
-class ResourceParamType(click.ParamType):
-    name = 'resource'
-
-    def __init__(self, nargs=-1, metavar='TEXT'):
-        self.nargs = nargs
-        self.metavar = metavar
-
-    def get_metavar(self, param):
-        metavar = self.metavar
-        if self.nargs == -1:
-            return '{0}[,{0},...]* | @filename'.format(metavar)
-        else:
-            return metavar
-
-    def convert(self, value, param, ctx):
-        def collect_resources(acc, resource_rep):
-            if resource_rep.startswith('@'):
-                for line in click.open_file(resource_rep[1:]):
-                    acc.append(line.strip())
-            else:
-                acc.append(resource_rep)
-            return acc
-        nargs = self.nargs
-        value = value.split(',')
-        resources = reduce(collect_resources, value, [])
-        if nargs > 0 and nargs != len(resources):
-            self.fail('Expected {} resources, but got {}'.format(nargs, len(resources)))
-        return resources
-
-    def __repr__(self):
-        'Resource(metavar={}, nargs={})'.fomat(self.metavar, self.nargs)
